requirements.txt
PyYAML==3.11
marisa-trie==0.7.2
psutil==3.4.1
urllib3==1.14
